added dfs

+++ models/red_ghost_model.py
@@ -5,13 +5,15 @@ import random
 import models.move_direction
 from appsettings import window_size
 from appsettings import image_store_path
+from path_finder import PathFinder
 
 pacman_model_directory = os.path.join(image_store_path, "red_ghost")
 frames_between_tile_pass = 16
 tile_switch_frame = 15
 
 class RedGhost(pygame.sprite.Sprite):
-    def __init__(self, spawn_tile, tile_size, player_model):
+    def __init__(self, spawn_tile, tile_size, player_model, player_input):
+        self.previous_frame_path_calc_player_tile = None
         self.player_model = player_model
         self.animation_tick_counter = 0
         self.current_sprite_index = 0
@@ -25,6 +27,7 @@ class RedGhost(pygame.sprite.Sprite):
         self.selected_move_direction = models.move_direction.right
         self.current_move_direction = models.move_direction.right
         self.stop = False
+        self.path_finder = PathFinder(player_input)
 
         pygame.sprite.Sprite.__init__(self)
 
@@ -33,10 +36,15 @@ class RedGhost(pygame.sprite.Sprite):
         self.set_pos(spawn_tile.rect.x, spawn_tile.rect.y)
 
     def update(self):
+        self.recalculate_path()
+
+        if self.path is None or len(self.path) == 0:
+            return
+        
         if(self.player_model.current_tile == self.current_tile):
             self.player_model.kill()
 
-        self.selected_move_direction = random.randint(0, 4)
+        self.selected_move_direction = self.path[0]
 
         self.animation_tick_counter += 1
         if self.animation_tick_counter % 3 == 0:
@@ -47,47 +55,42 @@ class RedGhost(pygame.sprite.Sprite):
     def move(self):
         if self.current_move_animation_frame == frames_between_tile_pass or self.stop:
             self.set_pos(self.current_tile.rect.x, self.current_tile.rect.y)
-            while True:
-                self.selected_move_direction = random.randint(0, 4)
-                if self.try_turn_in_direction(self.selected_move_direction):
-                    break
-            # if self.try_turn_in_direction(self.selected_move_direction) or self.try_turn_in_direction(self.current_move_direction):
-            #     self.stop = False
-            # else:
-            #     self.stop = True
+            self.try_turn_in_direction(self.selected_move_direction)
             self.current_move_animation_frame = 0
 
         if not self.stop:
             self.current_move_animation_frame += 1
             self.move_in_direction()
 
+    def recalculate_path(self):
+        if self.previous_frame_path_calc_player_tile is not None and self.previous_frame_path_calc_player_tile is self.player_model.current_tile:
+            return
+
+        self.path = self.path_finder.find(self.current_tile, self.player_model.current_tile).copy()
+
+        self.previous_frame_path_calc_player_tile = self.player_model.current_tile
+
     def move_in_direction(self):
-        if self.current_move_direction == models.move_direction.left:
+        self.move_sprite(self.current_move_direction)
+        for neigbor_info in self.current_tile.neighbours:
+            if neigbor_info.side is self.current_move_direction:
+                self.try_switch_tile(neigbor_info.tile)
+
+    def move_sprite(self, direction):
+        if direction == models.move_direction.left:
             self.move_x(-self.movement_speed)
-            self.try_switch_tile(self.current_tile.left_neighbor)
-        elif self.current_move_direction == models.move_direction.right:
+        elif direction == models.move_direction.right:
             self.move_x(self.movement_speed)
-            self.try_switch_tile(self.current_tile.right_neighbor)
-        elif self.current_move_direction == models.move_direction.up:
+        elif direction == models.move_direction.up:
             self.move_y(-self.movement_speed)
-            self.try_switch_tile(self.current_tile.upper_neighbor)
         else:
             self.move_y(self.movement_speed)
-            self.try_switch_tile(self.current_tile.bottom_neighbor)
 
     def try_turn_in_direction(self, direction):
-        if direction == models.move_direction.left and self.current_tile.left_neighbor.is_empty:
-            self.choose_direction(direction)
-            return True
-        elif direction == models.move_direction.right and self.current_tile.right_neighbor.is_empty:
-            self.choose_direction(direction)
-            return True
-        elif direction == models.move_direction.up and self.current_tile.upper_neighbor.is_empty:
-            self.choose_direction(direction)
-            return True
-        elif direction == models.move_direction.down and self.current_tile.bottom_neighbor.is_empty:
-            self.choose_direction(direction)
-            return True
+        for neigbor_info in self.current_tile.neighbours:
+            if neigbor_info.side == direction and neigbor_info.tile.is_empty:
+                self.choose_direction(direction)
+                return True
         return False
 
     def choose_direction(self, direction):
@@ -96,6 +99,7 @@ class RedGhost(pygame.sprite.Sprite):
     def try_switch_tile(self, neghbour_tile):
         if self.current_move_animation_frame == tile_switch_frame:
             self.current_tile = neghbour_tile
+            self.path.pop(0)
     
     def move_x(self, distance):
         self.rect.x += distance
